genaitor_server/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
flask-limiter
python-dotenv
requests